create subapp-v1-online-store
