--- samples/subapp1-online-store/src/02.main-body/subapp-manifest.js
@@ -1,6 +1,0 @@
-module.exports = {
-  type: "app",
-  name: "MainBody",
-  entry: "main-body",
-  serverEntry: "./server"
-};
