--- samples/subapp1-online-store/src/components/adv-grid.jsx
@@ -1,74 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { withStyles } from "@material-ui/core/styles";
-import GridList from "@material-ui/core/GridList";
-import GridListTile from "@material-ui/core/GridListTile";
-import GridListTileBar from "@material-ui/core/GridListTileBar";
-import IconButton from "@material-ui/core/IconButton";
-import StarBorderIcon from "@material-ui/icons/StarBorder";
-
-const styles = theme => ({
-  root: {
-    display: "flex",
-    flexWrap: "wrap",
-    justifyContent: "space-around",
-    overflow: "hidden",
-    backgroundColor: theme.palette.background.paper
-  },
-  gridList: {
-    width: 500,
-    height: 450,
-    // Promote the list into his own layer on Chrome. This cost memory but helps keeping high FPS.
-    transform: "translateZ(0)"
-  },
-  titleBar: {
-    background:
-      "linear-gradient(to bottom, rgba(0,0,0,0.7) 0%, " + "rgba(0,0,0,0.3) 70%, rgba(0,0,0,0) 100%)"
-  },
-  icon: {
-    color: "white"
-  }
-});
-
-function AdvancedGridList(props) {
-  const { classes, imagesData } = props;
-
-  return (
-    <div className={classes.root}>
-      <GridList cellHeight={200} spacing={1} className={classes.gridList}>
-        {imagesData.map(img => (
-          <GridListTile
-            key={img.contentUrl}
-            cols={img.featured ? 2 : 1}
-            rows={img.featured ? 2 : 1}
-          >
-            <img
-              width={img.thumbnail.width}
-              height={img.thumbnail.height}
-              src={img.thumbnailUrl}
-              alt={img.name}
-            />
-            <GridListTileBar
-              title={img.name}
-              titlePosition="top"
-              actionIcon={
-                <IconButton className={classes.icon}>
-                  <StarBorderIcon />
-                </IconButton>
-              }
-              actionPosition="left"
-              className={classes.titleBar}
-            />
-          </GridListTile>
-        ))}
-      </GridList>
-    </div>
-  );
-}
-
-AdvancedGridList.propTypes = {
-  classes: PropTypes.object.isRequired,
-  imagesData: PropTypes.array.isRequired
-};
-
-export default withStyles(styles)(AdvancedGridList);
