+++ samples/subapp1-online-store/src/main-body/reducers.jsx
